Revert "move package to top level"
This reverts commit dbd009c2723c5f0b7c76dbe73bda01ee7f5b9dcc.
